services/api/requirements.txt
fastapi[all]>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
psycopg[binary]>=3.1.16
geoalchemy2>=0.14.3
pydantic>=2.5.3
pydantic-settings>=2.1.0

# Auth
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt<4

# AWS
boto3>=1.34.25
aioboto3>=12.3.0

# AI/Copilot
langgraph>=0.0.20
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.10.0
anthropic>=0.18.0
google-generativeai>=0.3.0

# Utilities
redis>=5.0.1
circuitbreaker>=1.4.0
tenacity>=8.2.3
bleach>=6.1.0
python-multipart>=0.0.12

# Observability
structlog>=24.1.0
prometheus-client>=0.19.0
opentelemetry-api>=1.24.0
opentelemetry-sdk>=1.24.0
opentelemetry-exporter-otlp>=1.24.0
opentelemetry-instrumentation-fastapi>=0.45b0
opentelemetry-instrumentation-sqlalchemy>=0.45b0

# Security
slowapi>=0.1.9
bleach>=6.1.0

# Development
pytest>=7.4.4
pytest-cov>=4.1.0
pytest-asyncio>=0.23.3
hypothesis>=6.99.0
ruff>=0.1.14
mypy>=1.8.0
